refactor: move 3.8+ compat code to compat.py
